:building_construction: Rearrange codes of perfume detail data
