--- app/src/main/java/com/mashup/lastgarden/ui/main/PerfumeDetailData.kt
@@ -1,16 +1,0 @@
-package com.mashup.lastgarden.ui.main
-
-import com.mashup.lastgarden.data.vo.Tag
-
-data class PerfumeDetailData(
-    val perfumeDetailItems: List<PerfumeDetailItem>
-) {
-    data class PerfumeDetailItem(
-        val storyId: Int,
-        val thumbnailUrl: String? = null,
-        val userProfileImage: String? = null,
-        val userNickname: String,
-        val likeCount: Long? = 0L,
-        val tags: List<Tag>? = emptyList()
-    )
-}
