+++ app/src/main/java/com/mashup/lastgarden/ui/main/ScentListFragment.kt
@@ -53,7 +53,7 @@ class ScentListFragment : BaseViewModelFragment() {
             viewModel.perfumeDetailItems
                 .filterNotNull()
                 .collectLatest {
-                    perfumeDetailAdapter.submitList(it.perfumeDetailItems)
+                    perfumeDetailAdapter.submitList(it)
                 }
         }
     }
